chore(runtime): file-size and census gates for the audit commit
- shapes.rs: the census pins the parity asserts and the mint-then-stamp
  sabotage fixture to this file — return them, retarget the fixture at the
  _with_holes mint, and move the two cfg(test) keys-slot helpers to
  shapes_slot_list instead (keys_slot re-export split: owns_keys_slot is
  production code, descriptor_keys_slot is test-only).
- thread.rs: inline transfer_guard_tests module extracted to
  thread_transfer_guard_tests.rs (same #[path] pattern as its sibling).
- object/tests.rs: the three tombstone pins split into
  object/tombstone_tests.rs.

No behavior change; 2807 tests, fmt/census/class-id/file-size green.
